#61 moved GameJoinStatus
